+++ tksheet/find_window.py
@@ -9,7 +9,7 @@ from typing import Literal
 from .other_classes import (
     DotDict,
 )
-from .vars import (
+from .constants import (
     ctrl_key,
     rc_binding,
 )